Add show view for Job model
Create show view scaffolding using artisan make:view.

$ sail artisan make:view -- jobs.show

   INFO  View [resources/views/jobs/show.blade.php] created successfully.

Use the data extracted from the database through Job model and build the
view accordingly.

Let show action of the job controller to pass the data to the view.

Build the route for get request for auth and verified by middleware.
Build the name of jobs.show for the route.

Use Tailwind CSS to update app-layout for the job to look nice.