Add create view of Job model
Add create view scaffolding using artisan make:view.

$ sail artisan make:view -- jobs.create

   INFO  View [resources/views/jobs/create.blade.php] created successfully.

Build the form with two input boxes and colorful Save and Cancel buttons
for the markup of the form.

Build the create action of the job controller for the view.

Build the route for get request for auth and verified by middleware.
Build the name of jobs.create for the route.