Add individual job delete
Build job controller destroy action to delete the job and pass the job
deleted message to the index view by session.  Fix format of the view
that is out of alignment.

Add the delete button in the show view by the use of form.  Enhance the
edit button while there.

Build the route for the job delete.