Queue the mail when a new job created
Use Mail Facade to queue the mail when a new job is created.

Use public promoted job for easy serialization of the job.  The property
is public, making it accessible from outside the class.  Laravel's
serialization mechanism can properly access and restore the property.
The mailable works as expected

If we use private Job $job, the property is private, meaning it's only
accessible within the class.  When Laravel tries to serialize and later
unserialize the mailable class, the private property becomes
inaccessible during the unserialization process.  This will lead to the
failure of the mail.

Build job-posted view to use passed in job data.